+++ ci/proof.toml
@@ -166,6 +166,7 @@ name = "tokmd_core_ffi"
 kind = "rust"
 paths = [
   "crates/tokmd-core/src/lib.rs",
+  "crates/tokmd-core/src/receipts.rs",
   "crates/tokmd-core/src/ffi.rs",
   "crates/tokmd-core/src/ffi/**",
   "crates/tokmd-core/src/workflows/**",
